pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Michael Asumadu <michaelasumadu1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
python-dotenv = "^0.19.0"
Flask-SQLAlchemy = "^2.5.1"
flask-marshmallow = "^0.14.0"
Flask-Migrate = "^3.1.0"
marshmallow-sqlalchemy = "^0.26.1"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
